# specs/v1.5/swagger-saral-maintenance.yaml
swagger: "2.0"
info:
  description: "Store School information"
  version: "1.0.0"
  title: "SARAL APIs "
...
securityDefinitions:
    Bearer:
      type: apiKey
      name: Authorization
      in: header

    basicAuth:
      type: basic
schemes:
  - "https"
  - "http"
paths:
  /schools:
    get:
      tags:
      - School
      summary: "Get School List"
      description: "should return list of school"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: object
            properties:
              schools: 
                $ref: "#/definitions/Schools"
              
        "400":
          description: "unsuccessful operation"
  /schools/create:
    post:
      tags:
      - School
      summary: "create school"
      description: "create school"
      parameters:
      - name : body
        in: body
        schema:
          $ref: "#/definitions/CreateSchoolBody"
      responses:
        "201":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CreateSchoolResponse"
        "401":
          description: "schoolId: id already exist"
        "400":
          description: "unsuccessful operation"
  /schools/{schoolId}:
    delete:
      tags:
      - School
      summary: "Delete School By SchoolId"
      description: ""
      parameters:
        - in: path
          name: schoolId
          type: string
          required: true
      responses:
        "200": 
          description: "School has been deleted."
        "404":
          description: "School Id does not exist."
        "400":
          description: "unsuccessful operation"
      
    patch:
      tags:
      - School
      summary: "Update School By SchoolId"
      description: ""
      parameters:
        - in: path
          name: schoolId
          type: string
          required: true
        - in: body
          name: body
          schema:
            $ref: "#/definitions/UpdateSchoolBody"
      responses:
        "200": 
          description: "School has been updated."
        "404":
          description: "School Id does not exist."
        "400":
          description: "unsuccessful operation"
...
  /fetchStudentsByQuery:
    post:
      security:
        - Bearer: []
      tags:
      - Student
      summary: "Get students list"
      parameters:
       - in: body
         name: body
         schema:
            type: object
            required:
              - classId
            properties:
              classId:
                type: string
                example: "2"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/FetchStudentsByQueryResponse"
  /student:
    post:
      security:
        - Bearer: []
      tags:
      - Student
      summary: "Create Student"
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/CreateStudentBody"
      responses:
        "201":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CreateStudentResponse"
        "400":
          description: "Student Id is required."
  /student/{studentId}:
    delete:
      tags:
      - Student
      summary: "Delete Student "
      parameters:
        - in: path
          name: studentId
          type: string
          required: true
      responses:
        "200": 
          description: "Student has been deleted."
        "404":
          description: "Student Id does not exist"
        "400":
          description: "unsuccessful operation"
            

    patch:
      tags:
      - Student
      summary: "Update Student"
      parameters:
        - in: path
          name: studentId
          type: string
          required: true
        - in: body
          name: body
          schema:
            $ref: "#/definitions/UpdateStudentBody"
      responses:
        "200":
          description: "Student has been updated."
        "404":
          description: "Student Id does not exist."
        "400":
          description: "unsuccessful operation"
...
  /roi:
    post:
      security:
        - Bearer: []
      tags:
        - ROI
      summary: "Create ROI"
      parameters:
      - in: body
        name: body
        schema:
          $ref: "#/definitions/CreateROIBody"
      responses:
        "201":
          description: "successful operation"
          schema:
            $ref: "#/definitions/CreateROIResponse"
        "404":
          description: "examId does not exist"
        "400":
          description: "roiId already exist"
  /roi/{examId}:
    patch:
      security:
        - Bearer: []
      tags:
        - ROI
      summary: "Update ROI"
      parameters:
      - in: path
        name: examId
        type: number
        required: true
      - in: body
        name: body
        schema:
          $ref: "#/definitions/UpdateRoiBody"
      responses:
        "200":
          description: "ROI is updated successfully."
        "404":
          description: "ROI Id does not exist."
        "400":
          description: "unsuccessful operation"
            
    delete:
      security:
        - Bearer: []
      tags:
        - ROI
      summary: "Delete ROI"
      parameters:
      - in: path
        name: examId
        type: number
        required: true
      responses:
        "200":
          description: "ROI has been deleted successfully."
        "404":
          description: "ROI ID has been already deleted."
        "400":
          description: "unsuccessful operation"
...
definitions:
  Schools:
      type: array
      items:
        type: object
        properties:
          name:
            type: string
            example: "Dummy school 1"   
          schoolId:
            type: string
            example: "up001"
          state:
            type: string
            example: "up" 
          storeTrainingData:
            type: boolean
            default: false
            example: true
            description: "If this is set to true frontend application will return base 64 encoded images . In savesMarksAPI and getSavedScanAPI these APIs will get data and value of encoded images will be stored in these fields studentIdTrainingData and trainningData, studentIdTrainingData field will contain base 64 encoded images of rollnumber and trainningData field will contain base 64 encoded images for marks. "
  CreateSchoolBody:
    type: object
    required:
      - name
      - schoolId
      - password
      - state
    properties:
      name:
        type: string
        example: "dummy school 1"
      schoolId:
        type: string
        example: "u001"
      password: 
        type: string
        example: "abc"
      state:
        type: string
        example: "up"
      autoSync:
        type: boolean
        example: true
      autoSyncFrequency:
          type: number
          description: "Auto Sync frequency in milli-seconds. Saral App uses syncs data in local storage to backend based on this frequency value"
          example: 900000
      storeTrainingData:
        type: boolean
        default: false
        description: "If this is set to true frontend application will return base 64 encoded images . In savesMarksAPI and getSavedScanAPI these APIs will get data and value of encoded images will be stored in these fields studentIdTrainingData and trainningData, studentIdTrainingData field will contain base 64 encoded images of rollnumber and trainningData field will contain base 64 encoded images for marks. "
        example: true
      tags:
        type: boolean
        example: true
      autoSyncBatchSize:
        type: number
        description:  "Limit of student records to be stored in local storage before we push to backend."
        example: 10
  CreateSchoolResponse:
    type: object
    properties:
      schools:
        type: object
        properties:
          name: 
            type: string
            example: "dummy school 1"
          schoolId:
            type: string
            example: "u001"
          state:
            type: string
            example: "up"
      token:
        type: object
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MTMwOWQyYjJlNzEyMzAyZGY5NmM3NjEiLCJpYXQiOjE2MzA1NzU5MTV9.FCt2Q4R_sN1ctdOfoOPzBzfdUohAPA-TlLFtJJEeTBE"
  UpdateSchoolBody:
    type: object
    properties:
      name:
        type: string
        example: "Dummy school 1"
      state:
        type: string
        example: "up"
      udisceCode:
        type: string
        example: "1"
      autoSync:
        type: boolean
        example: true
      autoSyncFrequency:
        type: number
        description: "Auto Sync frequency in milli-seconds. Saral App uses syncs data in local storage to backend based on this frequency value"
        example: 900000
      storeTrainingData:  
        type: boolean
        default: false
        example: true
        description: "If this is set to true frontend application will return base 64 encoded images . In savesMarksAPI and getSavedScanAPI these APIs will get data and value of encoded images will be stored in these fields studentIdTrainingData and trainningData, studentIdTrainingData field will contain base 64 encoded images of rollnumber and trainningData field will contain base 64 encoded images for marks. "
      tags:
        type: boolean
        example: false
      autoSyncBatchSize:
        type: number
        description:  "Limit of student records to be stored in local storage before we push to backend."
        example: 10
  UpdateStudentBody:
    type: object
    required:
      - name
    properties:
      name: 
        type: string
        example: "1210001"
  CreateStudentBody:
    type: object
    required:
       - name
       - studentId
    properties:
      name:
        type: string
        description: "Student Name"
        example: "ram"
      section:
        type: string
        description: "Class Section"
        example: "A"
      studentId:
        type: string
        description: "Student ID"
        example: "1210001"
      studentClass:
        $ref: '#/definitions/StudentClass'
  FetchStudentsByQueryResponse:
    type: object
    properties:
      studentClass:
        $ref: '#/definitions/StudentClassName'
      section:
        type: string
        example: "D"
      name:
        type: string
        example: "2"
      studentId:
        type: string
        example: "1210001"
      schoolId:
        type: string
        example: "u001"
  CreateStudentResponse:
    type: object
    properties:
      studentClass: 
        $ref: '#/definitions/StudentClassName' 
      section:
        type: string
        description: "Class Section"
        example: "D"
      name:
        type: string
        description: "Student Name"
        example: "Ram"
      studentId:
        type: string
        description: "Student Id"
        example: "1210001"
      schoolId:
        type: string
        description: "School Id"
        example: "u001"
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  StudentClass:
    type: array
    items:
      type: object
      properties: 
        classId:
          type: string
          description: "Id of class"
          example: "2"
  StudentClassName:
    type: array
    items:
      type: object
      properties: 
        classId:
          type: string
          description: "Id of class"
          example: "2"
        className:
          type: string
          description: "class Name"
          example: "Class-2"
...
  rect:
    type: object
    properties:
      top:
        type: string
        example: "129"
      left:
        type: string
        example: "56"
      bottom:
        type: string
        example: "73"
      right:
        type: string
        example: "181"
  roi:
    type: array
    items:
      type: object
      properties:
        annotationTags:
          type: string
          example: "ROLLNUMBERID1"
        extractionMethod:
          type: string
          enum: ["NUMERIC_CLASSIFICATION", "CELL_OMR"]
          example: "NUMERIC_CLASSIFICATION"
        roiId:
          type: number
          example: 1
          description: RoiId field is global across school or organisation.
        index:
          type: string
          example: 0
        rect:
          $ref: '#/definitions/rect'
  render:
    type: object
    properties:
      index:
        type: string
        example: 1
  format:
    type: object
    properties:
      name:
        type: string
        example: "ROLLNUMBER"
      value:
        type: string
        example: "ROLLNUMBER"
  validate:
    type: object
    properties:
      regExp:
        type: string
        example: ""
  cells:
    type: array
    items:
      type: object
      properties:
        cellId:
          type: number
          example: 1
        rois:
          $ref: '#/definitions/roi'
        render:
          $ref: '#/definitions/render'
        format:
          $ref: '#/definitions/format'
        validate:
          $ref: '#/definitions/validate'
  rois:
    type: object
    properties:
      layout:
        type: object
        properties:
          version: 
            type: string
            example: "1.0"
          name:
            type: string
            example: "UP SAT Exam Sheet Form"
          identifierPrefix:
            type: string
            example: ROLLNUMBERID
          threshold:
            type: object
            properties:
              minWidth:
                type: number
                example: 690
                description: Minimum widht of layout for detection
              minHeight:
                type: number
                example: 390
                description: Minimum height of layout for detection
              detectionRadius:
                type: number
                example: 12
              experimentalOMRDetection:
                type: boolean
                example: true
                description: when this flag is true pencil detection of layout will be predicted .
          cells:
            $ref: '#/definitions/cells' 
  CreateROIBody:
    type: object
    required:
      - subject
      - classId
      - type
      - roi
    properties:
      subject:
        type: string
        example: "Maths"
      classId:
        type: string
        example: "2"
      set:
        type: string
        example: A
      type: 
        type: string
        example: "uphindi1s"
      roi:
        $ref: '#/definitions/rois'
  CreateROIResponse:
    type: object
    properties:
      roiId:
        type: number
        example: 1
        description: RoiId field is global across school or organisation.
      classId:
        type: string
        example: "2"
      subject:
        type: string
        example: "Maths"
      state:
        type: string
        example: "up"
      createdAt:
        type: string
        format: date-time
  UpdateRoiBody:
    type: object
    required:
      - roi
    properties:
      roi:
        $ref: "#/definitions/rois"
...